requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
-e file:.
    # via -r requirements.in
astor==0.8.1
    # via flake8-qgis
atomicwrites==1.4.0
    # via
    #   -r requirements.in
    #   pytest
attrs==22.1.0
    # via pytest
certifi==2022.9.24
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
colorama==0.4.5
    # via pytest
coverage[toml]==7.2.7
    # via pytest-cov
distlib==0.3.6
    # via virtualenv
filelock==3.8.0
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements.in
    #   flake8-pyproject
    #   flake8-qgis
    #   flake8-spellcheck
flake8-pyproject==1.2.3
    # via -r requirements.in
flake8-qgis==1.0.0
    # via -r requirements.in
flake8-spellcheck==0.28.0
    # via -r requirements.in
identify==2.5.6
    # via pre-commit
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via qgis-plugin-dev-tools
iniconfig==1.1.1
    # via pytest
mccabe==0.7.0
    # via flake8
mypy==1.13.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.7.0
    # via pre-commit
packaging==23.2
    # via
    #   pytest
    #   qgis-plugin-dev-tools
platformdirs==2.5.2
    # via virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==3.4.0
    # via -r requirements.in
py==1.11.0
    # via pytest
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyqt5-stubs==5.15.6.0
    # via -r requirements.in
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-dotenv
    #   pytest-mock
    #   pytest-order
    #   pytest-qgis
    #   pytest-qt
    #   pytest-timeout
pytest-cov==2.12.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-mock==3.7.0
    # via -r requirements.in
pytest-order==1.0.0
    # via -r requirements.in
pytest-qgis==1.3.0
    # via -r requirements.in
pytest-qt==3.3.0
    # via -r requirements.in
pytest-timeout==1.4.2
    # via -r requirements.in
python-dotenv==0.21.0
    # via
    #   pytest-dotenv
    #   qgis-plugin-dev-tools
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   pre-commit
qgis-plugin-dev-tools==0.5.0
    # via -r requirements.in
qgis-plugin-tools==0.2.0
    # via segment-reshape-qgis-plugin
requests==2.28.1
    # via qgis-plugin-dev-tools
ruff==0.8.1
    # via -r requirements.in
toml==0.10.2
    # via pytest
tomli==2.0.1
    # via qgis-plugin-dev-tools
typing-extensions==4.12.2
    # via mypy
urllib3==1.26.12
    # via requests
virtualenv==20.16.5
    # via pre-commit
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
